+++ R/backend-oracle.R
@@ -54,7 +54,7 @@ sql_translate_env.Oracle <- function(con) {
 }
 
 #' @export
-sql_explain.Oracle <- function(con, sql, ...) {
+sql_query_explain.Oracle <- function(con, sql, ...) {
   build_sql(
     "EXPLAIN PLAN FOR ", sql, ";\n",
     "SELECT PLAN_TABLE_OUTPUT FROM TABLE(DBMS_XPLAN.DISPLAY()));",
@@ -63,7 +63,7 @@ sql_explain.Oracle <- function(con, sql, ...) {
 }
 
 #' @export
-sql_analyze.Oracle <- function(con, table, ...) {
+sql_table_analyze.Oracle <- function(con, table, ...) {
   # https://docs.oracle.com/cd/B19306_01/server.102/b14200/statements_4005.htm
   build_sql("ANALYZE TABLE ", as.sql(table), " COMPUTE STATISTICS", con = con)
 }
@@ -94,7 +94,7 @@ sql_translate_env.OraConnection <- sql_translate_env.Oracle
 sql_select.OraConnection <- sql_select.Oracle
 
 #' @export
-sql_analyze.OraConnection <- sql_analyze.Oracle
+sql_table_analyze.OraConnection <- sql_table_analyze.Oracle
 
 #' @export
 sql_subquery.OraConnection <- sql_subquery.Oracle