+++ R/backend-teradata.R
@@ -85,7 +85,7 @@ sql_translate_env.Teradata <- function(con) {
   )}
 
 #' @export
-sql_analyze.Teradata <- function(con, table, ...) {
+sql_table_analyze.Teradata <- function(con, table, ...) {
   # https://www.tutorialspoint.com/teradata/teradata_statistics.htm
   build_sql("COLLECT STATISTICS ", as.sql(table) , con = con)
 }